requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
decorator==4.4.2
dnspython==1.16.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
odnoklassniki==1.0
pycparser==2.20
pymongo==3.10.1
pyOpenSSL==19.1.0
python-telegram-bot==12.6.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
